% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data from City of Windsor Open Data Catalogue}
\usage{
get_data(data)
}
\arguments{
\item{data}{Name of one or more datasets. A character}
}
\value{
A list containing one or more datasets
}
\description{
Get data from City of Windsor Open Data Catalogue
}
\examples{
get_data(c("AbandonedVehicle_YTD", "3DayParkingInfraction_YTD"))
}
